[FIX] clean file causing too many events

+++ app/util.py
@@ -1,5 +1,6 @@
 import multiprocessing
 import shutil
+import os
 
 from django.http import *
 from django.utils import timezone
@@ -60,14 +61,19 @@ def clean():
     current = timezone.now()
 
     for task in Task.objects.all():
+        dirname = task.get_dirname()
+        if not os.path.isdir(dirname):
+            continue
         if (current - task.start_time).total_seconds() >= 24 * 3600:
-            dirname = task.get_dirname()
             Event(event_type="clean_task", event_message="id=%d" % task.task_id).save()
             shutil.rmtree(dirname)
 
     for mania_file in ManiaFile.objects.all():
+        dirname = mania_file.get_dirname()
+        if not os.path.isdir(dirname):
+            continue
         if (current - mania_file.save_time).total_seconds() >= 24 * 3600:
-            shutil.rmtree(mania_file.get_dirname())
+            shutil.rmtree(dirname)
             Event(event_type="clean_file",
                   event_message="id=%d, type=%s" % (
                   mania_file.file_id, mania_file.file_type)).save()
